requirements.txt
Flask
flask-cors
scikit-learn
python-dotenv